requirements.txt
pyrogram==2.0.106
gspread==6.0.0
python-dotenv==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
asyncio==3.4.3
aiofiles==23.2.0